OCR_TO_EXCEL_APP/requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
tkintertable>=1.3.2
python-docx>=1.1.0
pdf2image>=1.16.3
pyautogui>=0.9.54